[move] : textInput 관련 폴더끼리 묶은 뒤 elements에서 한 단계 상위 레벨로 이동
