+++ apps/storybook/src/stories/shared/components/textInput/textInputRightAddOn/rightAddOnArrowIcon.stories.tsx
@@ -1,4 +1,4 @@
-import { RightAddOnArrowIcon } from '@shared/components/elements/textInputRightAddOn';
+import { RightAddOnArrowIcon } from '@repo/shared/src/components/textInput/textInputRightAddOn';
 import { Meta, StoryFn } from '@storybook/react';
 
 export default {
