+++ apps/storybook/src/stories/shared/components/textInput/textInputRightAddOn/rightAddOnCalendarIcon.stories.tsx
@@ -1,4 +1,4 @@
-import { RightAddOnCalendarIcon } from '@shared/components/elements/textInputRightAddOn';
+import { RightAddOnCalendarIcon } from '@repo/shared/src/components/textInput/textInputRightAddOn';
 import { Meta, StoryFn } from '@storybook/react';
 
 export default {
