+++ apps/storybook/src/stories/shared/components/textInput/textInputRightAddOn/rightAddOnTimeDisplay.stories.tsx
@@ -1,4 +1,4 @@
-import { RightAddOnTimeDisplay } from '@shared/components/elements/textInputRightAddOn';
+import { RightAddOnTimeDisplay } from '@repo/shared/src/components/textInput/textInputRightAddOn';
 import { Meta, StoryFn } from '@storybook/react';
 
 export default {
