+++ packages/shared/src/components/textInput/textInputLeftAddOn/LeftAddOnSearchIcon.tsx
