+++ packages/shared/src/components/textInput/textInputLeftAddOn/index.ts
