+++ packages/shared/src/components/textInput/textInputRightAddOn/RightAddOnArrowIcon.tsx
