+++ packages/shared/src/components/textInput/textInputRightAddOn/RightAddOnCalendarIcon.tsx
