+++ packages/shared/src/components/textInput/textInputRightAddOn/RightAddOnTimeDisplay.tsx
