+++ packages/shared/src/components/textInput/textInputRightAddOn/index.ts
